create components container and ui for taskitem; move completion change and single task deletion to task item container
