+++ src/components/TaskListUI.tsx
@@ -1,11 +1,11 @@
+import { Dispatch, SetStateAction } from 'react';
 import { Task } from '../utils/types';
-import TaskItem from './TaskItem';
+import TaskItemContainer from './TaskItemContainer';
 
 interface TaskListUIProps {
   uncompletedTasks: Task[];
   completedTasks: Task[];
-  handleTaskCompletionChange: (taskId: string) => Promise<void>;
-  handleSingleTaskDelete: (taskId: string) => Promise<void>;
+  setTasks: Dispatch<SetStateAction<Task[]>>;
 }
 
 const COMPLETED_TASKS = 10; //number of completed tasks to show
@@ -13,8 +13,7 @@ const COMPLETED_TASKS = 10; //number of completed tasks to show
 function TaskListUI({
   uncompletedTasks,
   completedTasks,
-  handleTaskCompletionChange,
-  handleSingleTaskDelete,
+  setTasks,
 }: TaskListUIProps) {
   return (
     <>
@@ -23,11 +22,7 @@ function TaskListUI({
         <ul>
           {uncompletedTasks.map((task) => (
             <li key={task._id}>
-              <TaskItem
-                task={task}
-                handleTaskCompletionChange={handleTaskCompletionChange}
-                handleSingleTaskDelete={handleSingleTaskDelete}
-              />
+              <TaskItemContainer task={task} setTasks={setTasks} />
             </li>
           ))}
         </ul>
@@ -37,11 +32,7 @@ function TaskListUI({
         <ul>
           {completedTasks.slice(0, COMPLETED_TASKS).map((task) => (
             <li key={task._id}>
-              <TaskItem
-                task={task}
-                handleTaskCompletionChange={handleTaskCompletionChange}
-                handleSingleTaskDelete={handleSingleTaskDelete}
-              />
+              <TaskItemContainer task={task} setTasks={setTasks} />
             </li>
           ))}
           {completedTasks.length > COMPLETED_TASKS && <li>...</li>}
